fix: generate unique ids for new images
There was an issue where after regenerating the image, the old image
still keeps on showing up.

This was because regenerated images simply replace their old images
while keeping the id the same. By using the same id, the img url stayed
the same, and so the device kept on fetching the image from the cache
instead of refetching from the backend.

I've tried multiple ways to disable cache, but none of them seemed to
work.

So, the easiest way was just to generate new ids for the new images and
delete the old images when regenerating.