feat: remove uneeded book chapter, title, and author handling
We don't need to do the whole tocFlatten thing to get the current
chapter. We can just get the current section and get the chapter label
from there

For book title and author, the function `visualizeHighlight` in our
backend service will handle fetching those details from the database.
So, we don't need to get those details in the book reader page.