refactor: rename "annotation" to "data"
I just feel this is a more appropriate name because we are getting data
from the database to populate the Annotation list

Who knows, maybe the data in the database will change and "annotation"
won't be a proper name for it.

Best to just use a general name like "data" for this case.